requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
praw==7.7.1
requests==2.31.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.3
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
